requirements.txt
django==5.0.6
django-ckeditor==6.7.1